% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_extract.R
\name{cycle_extract}
\alias{cycle_extract}
\title{Extract cycles into a better format from ripsgen() function}
\usage{
cycle_extract(n = 30, dim = 2, diagram = NA, cycles = NA)
}
\arguments{
\item{n}{Number of points to generate (required if "dat" is not supplied).}

\item{dim}{The underlying ambient dimension of the result, i.e., dim=2 yields
data in a a unit square (required if "dat" is not supplied).}

\item{diagram}{The Rips diagram as returned by ripsgen()$diagram.}

\item{cycles}{The representative cycles as returned by ripsgen()$cycles.}
}
\value{
something here
}
\description{
Extract cycles into a better format from ripsgen() function
}
\examples{
2+2
}
