% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycs.R
\docType{data}
\name{cycs}
\alias{cycs}
\title{Cycle Line Segments}
\format{
A data frame with 70605 rows and 7 columns
\describe{
  \item{run}{Dataset Identifier}
  \item{npts}{Number of points in the dataset}
  \item{cycle_ix}{Identifies the representative cycles this line segment belongs to.}
  \item{x0}{Initial x value}
  \item{y0}{Initial y value}
  \item{x1}{Final x value}
  \item{y1}{Final y value}
}
}
\usage{
cycs
}
\description{
Line segments forming 363 cycles from 297 different datasets ranging from 3 to 300 points.
To be used in conjuction with "diags" and "pnts" datasets.
}
\keyword{datasets}
