% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagen.R
\name{datagen}
\alias{datagen}
\title{Generate randomly distributed data in an n-Cube.}
\usage{
datagen(n = 30, dim = 2)
}
\arguments{
\item{n}{Number of points to generate}

\item{dim}{The underlying ambient dimension of the result, i.e., dim=2 yields
data in a a unit square}
}
\value{
A tibble with "n" rows and "dim" columns of random data.
}
\description{
Generate randomly distributed data in an n-Cube.
}
\examples{
dat <- datagen(50, 3)
pairs(dat)
}
