% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diags.R
\docType{data}
\name{diags}
\alias{diags}
\title{Rips Diagrams}
\format{
A data frame with 54802 rows and 6 columns
\describe{
  \item{run}{Dataset Identifier}
  \item{cycle_ix}{Identifier for a homology class}
  \item{npts}{Number of points in the dataset}
  \item{dim}{Homological Dimension}
  \item{birth}{Initial epsilon value}
  \item{death}{Initial epsilon value}
}
}
\usage{
diags
}
\description{
Rips diagrams from 297 different datasets ranging from 3 to 300 points.
To be used in conjuction with "pnts" and "cycs" datasets.
}
\keyword{datasets}
