% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cycle.R
\name{plot_cycle}
\alias{plot_cycle}
\title{Plot a cycle on the supplied data given by the cycle segments}
\usage{
plot_cycle(data, ordered_cycles, plot_data = TRUE)
}
\arguments{
\item{data}{Random data from the experiment, must be only data coordinates
(i.e., "x", "y", "z") along with an id for the cycle as the first column.}

\item{ordered_cycles}{The ordered cycles provided by cycle_extract().}

\item{plot_data}{Boolean to toggle plotting underlying data}
}
\value{
A plot
}
\description{
Plot a cycle on the supplied data given by the cycle segments
}
\examples{
require(dplyr)
rr <- ripsgen()
plot_cycle(
  data = rr$data,
  ordered_cycles = cycle_extract(
    diagram = rr$diagram,
    cycle_segments = rr$cycle_segments
  )
)

# you can also plot one cycle at a time if you loop the following appropriately
cycle <- cycle_extract(
  diagram = rr$diagram,
  cycle_segments = rr$cycle_segments
) \%>\% filter(cycle_ix == 31)

}
