% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cycle.R
\name{plot_cycle_raw}
\alias{plot_cycle_raw}
\title{Plot a cycle on the supplied data given by the cycle segments}
\usage{
plot_cycle_raw(data = NA, cycle_segments = NA)
}
\arguments{
\item{data}{Random data from the experiment}

\item{cycle_segments}{The provided cycle segments, looks like "index", "x0", "y0", "x1", "y1".}
}
\value{
A plot
}
\description{
Plot a cycle on the supplied data given by the cycle segments
}
\examples{
2+2
}
