% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cycle_segments.R
\name{plot_cycle_segments}
\alias{plot_cycle_segments}
\title{Plot a cycle on the supplied data given by the cycle segments}
\usage{
plot_cycle_segments(data, cycle_segments)
}
\arguments{
\item{data}{Random data from the experiment, must be only data coordinates (i.e., "x", "y", "z")}

\item{cycle_segments}{The provided cycle segments, looks like "index", "x0", "y0", "x1", "y1".}
}
\value{
A plot
}
\description{
Plot a cycle on the supplied data given by the cycle segments
}
\examples{
rr <- ripsgen()
plot_cycle_segments(data = rr$data, cycle_segments = rr$cycle_segments)
}
