% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnts.R
\docType{data}
\name{pnts}
\alias{pnts}
\title{Random Point Cloud}
\format{
A data frame with 45147 rows and 4 columns
\describe{
  \item{run}{Dataset identifier}
  \item{npts}{Number of points in the dataset}
  \item{x}{x-coordinate}
  \item{y}{y-coordinate}
}
}
\usage{
pnts
}
\description{
Random point cloud data from 297 different runs of data, ranging from 3 to 300 points.
To be used in conjuction with "diags" and "cycs" datasets.
}
\keyword{datasets}
