% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripsgen.R
\name{ripsgen}
\alias{ripsgen}
\title{Generate Rips Complex for data}
\usage{
ripsgen(n = 30, dim = 2, dat)
}
\arguments{
\item{n}{Number of points to generate (required if "dat" is not supplied).}

\item{dim}{The underlying ambient dimension of the result, i.e., dim=2 yields
data in a a unit square (required if "dat" is not supplied).}

\item{dat}{An optional tibble, dataframe, or matrix, with columns of randomly distributed data only (i.e., "x", "y", and "z" being all random uniform variables).}
}
\value{
A List with the original or randomly generated data, along with the Rips diagram, birth and death times, and cycles segments that can draw the cycles on the random data.
}
\description{
Generate Rips Complex for data
}
\examples{
rr <- ripsgen(50, 2)
rr$diagram
}
