% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_edge.R
\name{select_edge}
\alias{select_edge}
\title{Find the edge to shrink or split a representative cycle}
\usage{
select_edge(ordered_cycle)
}
\arguments{
\item{ordered_cycle}{Must be only data coordinates (i.e., "x", "y", "z")
in a sequential (adjacent) order.}
}
\value{
List of the adjacency matrix, the indices of the edge to split or
shrink upon, and the epsilon value this occurs at.
}
\description{
Find the edge to shrink or split a representative cycle
}
\examples{
2 + 2 # TODO
}
