% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cycle.R
\name{plot_cycle}
\alias{plot_cycle}
\title{Plot a cycle on the supplied data given by the cycle segments}
\usage{
plot_cycle(data, ordered_cycles)
}
\arguments{
\item{data}{Random data from the experiment, must be only data coordinates (i.e., "x", "y", "z")}

\item{ordered_cycles}{The ordered cycles provided by cycle_extract().}
}
\value{
A plot
}
\description{
Plot a cycle on the supplied data given by the cycle segments
}
\examples{
rr <- ripsgen()
plot_cycle(
  data = rr$data,
  ordered_cycles = cycle_extract(diagram = rr$diagram, cycles = rr$cycle_segments)
)
}
